Atualiza configuração e adiciona testes ao projeto
- Remove configuração de depuração antiga e adiciona nova no `launch.json`.
- Atualiza caminhos no `tasks.json` para refletir nova estrutura.
- Modifica GUIDs dos projetos no `03-api-book-rating-manager.sln` e adiciona projeto de testes.
- Refatora `ExceptionHandlingMiddleware.cs` para incluir tratamento de exceções.
- Adiciona verificação de ID no `DeleteBookCommandHandler.cs` para evitar exclusões inválidas.
- Atualiza `Book-Evaluation-Management-System.Tests.csproj` com novas referências e configurações.
- Implementa testes em `DeleteBookCommandTests.cs` para validação de exclusão.
- Adiciona testes em `EditBookCommandTest.cs` para verificação de edição.
- Cria testes em `RegisterNewBookCommandTests.cs` para validação de adição de livros e tratamento de erros.

+++ Book-Evaluation-Management-System.Application/Commands/Book/DeleteBook/DeleteBookCommandHandler.cs
@@ -12,6 +12,11 @@ namespace Book_Evaluation_Management_System.Application.Commands.Book.DeleteBook
         }
         public async Task<Unit> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
         {
+
+            if (request.Id == 0)
+            {
+                throw new ArgumentException("Id cannot be empty.", nameof(request.Id));
+            }
             await _bookRepository.DeleteBookAsync(request.Id);
 
             return Unit.Value;
